# Conflicts:
#	src/main/java/gov/ca/cwds/rest/services/ParticipantService.java

+++ src/main/java/gov/ca/cwds/data/cms/CmsDocumentDao.java
@@ -1,5 +1,7 @@
 package gov.ca.cwds.data.cms;
 
+import static java.lang.Math.min;
+
 import java.io.File;
 import java.io.FileOutputStream;
 import java.io.IOException;
@@ -8,9 +10,10 @@ import java.nio.file.Paths;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
+import java.util.function.Supplier;
+
 import javax.xml.bind.DatatypeConverter;
 
-import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.http.util.ByteArrayBuffer;
 import org.hibernate.SessionFactory;
@@ -19,6 +22,7 @@ import org.slf4j.LoggerFactory;
 
 import com.google.inject.Inject;
 
+import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
 import gov.ca.cwds.data.BaseDaoImpl;
 import gov.ca.cwds.data.persistence.cms.CmsDocument;
 import gov.ca.cwds.data.persistence.cms.CmsDocumentBlobSegment;
@@ -27,8 +31,6 @@ import gov.ca.cwds.rest.filters.RequestExecutionContext;
 import gov.ca.cwds.rest.services.ServiceException;
 import gov.ca.cwds.rest.util.jni.CmsPKCompressor;
 import gov.ca.cwds.rest.util.jni.LZWEncoder;
-
-import static java.lang.Math.min;
 
 /**
  * Data Access Object (DAO) for legacy, compressed CMS documents.
@@ -47,6 +49,8 @@ public class CmsDocumentDao extends BaseDaoImpl<CmsDocument> {
   public static final String COMPRESSION_TYPE_PK_FULL = "PKWare02";
   public static final String COMPRESSION_TYPE_PLAIN_FULL = "PLAIN_00";
   public static final int BLOB_SEGMENT_LENGTH = 4000;
+
+  protected Supplier<LZWEncoder> lzwSupplier = LZWEncoder::new;
 
   /**
    * Constructor.
@@ -78,7 +82,7 @@ public class CmsDocumentDao extends BaseDaoImpl<CmsDocument> {
         final String sequence = StringUtils.leftPad(String.valueOf(++i), 4, '0');
         final int segmentLength = min(compressed.length - segmentStart, BLOB_SEGMENT_LENGTH);
         blobs.add(new CmsDocumentBlobSegment(doc.getId(), sequence,
-            Arrays.copyOfRange(compressed,segmentStart, segmentStart + segmentLength)));
+            Arrays.copyOfRange(compressed, segmentStart, segmentStart + segmentLength)));
         segmentStart += segmentLength;
       }
 
@@ -91,12 +95,12 @@ public class CmsDocumentDao extends BaseDaoImpl<CmsDocument> {
       doc.setLastUpdatedId(StringUtils.isNotBlank(ctx.getStaffId()) ? ctx.getStaffId() : "0x5");
 
     } catch (Exception e) {
-      LOGGER.error("ERROR COMPRESSING PK! {}", e.getMessage());
-      throw new ServiceException("ERROR COMPRESSING PK! " + e.getMessage(), e);
+      errorCompressing(e);
     }
 
     return blobs;
   }
+
   /**
    * No compressed document blob segments.
    *
@@ -104,7 +108,7 @@ public class CmsDocumentDao extends BaseDaoImpl<CmsDocument> {
    * @param base64 base64 encoded bytes
    * @return new blobs in order
    */
-  private List<CmsDocumentBlobSegment> compressPLAIN(final CmsDocument doc, String base64) {
+  protected List<CmsDocumentBlobSegment> compressPlain(final CmsDocument doc, String base64) {
     final List<CmsDocumentBlobSegment> blobs = new ArrayList<>();
     try {
       final byte[] plain = DatatypeConverter.parseBase64Binary(base64);
@@ -114,8 +118,10 @@ public class CmsDocumentDao extends BaseDaoImpl<CmsDocument> {
       while (segmentStart < plain.length) {
         final String sequence = StringUtils.leftPad(String.valueOf(++i), 4, '0');
         final int segmentLength = min(plain.length - segmentStart, BLOB_SEGMENT_LENGTH);
-        blobs.add(new CmsDocumentBlobSegment(doc.getId(), sequence,
-                Arrays.copyOfRange(plain,segmentStart, segmentStart + segmentLength)));
+        final CmsDocumentBlobSegment blob = new CmsDocumentBlobSegment(doc.getId(), sequence,
+            Arrays.copyOfRange(plain, segmentStart, segmentStart + segmentLength));
+        blobs.add(blob);
+        doc.addBlobSegment(blob);
         segmentStart += segmentLength;
       }
 
@@ -128,8 +134,7 @@ public class CmsDocumentDao extends BaseDaoImpl<CmsDocument> {
       doc.setLastUpdatedId(StringUtils.isNotBlank(ctx.getStaffId()) ? ctx.getStaffId() : "0x5");
 
     } catch (Exception e) {
-      LOGGER.error("ERROR COMPRESSING PLAIN! {}", e.getMessage());
-      throw new ServiceException("ERROR COMPRESSING PLAIN! " + e.getMessage(), e);
+      errorCompressing(e);
     }
 
     return blobs;
@@ -146,26 +151,26 @@ public class CmsDocumentDao extends BaseDaoImpl<CmsDocument> {
     String retval = "";
 
     if (doc.getCompressionMethod().endsWith(COMPRESSION_TYPE_LZW)) {
-      LZWEncoder lzw = new LZWEncoder();
+      final LZWEncoder lzw = lzwSupplier.get();
       if (!lzw.didLibraryLoad()) {
-        LOGGER.warn("LZW compression not enabled!");
+        LOGGER.warn("LZW COMPRESSION NOT ENABLED!");
       } else {
         retval = decompressLZW(doc);
       }
     } else if (doc.getCompressionMethod().endsWith(COMPRESSION_TYPE_PK)) {
       retval = decompressPK(doc);
     } else if (doc.getCompressionMethod().endsWith(COMPRESSION_TYPE_PLAIN)) {
-      retval = decompressPLAIN(doc);
+      retval = decompressPlain(doc);
     } else {
       LOGGER.error("UNSUPPORTED COMPRESSION METHOD! {}", doc.getCompressionMethod());
     }
 
     return retval;
   }
+
   /**
    * Compress (deflate) a document by determining the compression type, calling appropriate library
    * and splitting on blob segments.
-   *
    *
    * @param doc document to compress
    * @param base64 base64 encoded bytes
@@ -175,16 +180,16 @@ public class CmsDocumentDao extends BaseDaoImpl<CmsDocument> {
     List<CmsDocumentBlobSegment> retval = null;
 
     if (doc.getCompressionMethod().endsWith(COMPRESSION_TYPE_LZW)) {
-      LZWEncoder lzw = new LZWEncoder();
+      final LZWEncoder lzw = lzwSupplier.get();
       if (!lzw.didLibraryLoad()) {
-        LOGGER.warn("LZW compression not enabled!");
+        LOGGER.warn("LZW COMPRESSION NOT ENABLED!");
       } else {
         retval = compressLZW(doc, base64);
       }
     } else if (doc.getCompressionMethod().endsWith(COMPRESSION_TYPE_PK)) {
       retval = compressPK(doc, base64);
     } else if (doc.getCompressionMethod().endsWith(COMPRESSION_TYPE_PLAIN)) {
-      retval = compressPLAIN(doc, base64);
+      retval = compressPlain(doc, base64);
     } else {
       LOGGER.error("UNSUPPORTED COMPRESSION METHOD! {}", doc.getCompressionMethod());
     }
@@ -209,20 +214,19 @@ public class CmsDocumentDao extends BaseDaoImpl<CmsDocument> {
     try {
       final ByteArrayBuffer buf = new ByteArrayBuffer(doc.getDocLength().intValue());
       for (CmsDocumentBlobSegment seg : doc.getBlobSegments()) {
-        buf.append(seg.getDocBlob(), 0, seg.getDocBlob().length);
+        final byte[] blob = seg.getDocBlob();
+        buf.append(blob, 0, blob.length);
       }
 
       final byte[] bytes = new CmsPKCompressor().decompressBytes(buf.buffer());
       LOGGER.debug("DAO: bytes len={}", bytes.length);
       retval = DatatypeConverter.printBase64Binary(bytes);
     } catch (Exception e) {
-      LOGGER.error("ERROR DECOMPRESSING PK! {}", e.getMessage());
-      throw new ServiceException("ERROR DECOMPRESSING PK! " + e.getMessage(), e);
+      errorDecompressing(e);
     }
 
     return retval;
   }
-
 
   /**
    * Decompress (inflate) a non-compressed document by assembling blob segments
@@ -234,21 +238,22 @@ public class CmsDocumentDao extends BaseDaoImpl<CmsDocument> {
    * @param doc PLAIN archive to decompress
    * @return base64-encoded String of decompressed document
    */
-  private String decompressPLAIN(CmsDocument doc) {
+  protected String decompressPlain(CmsDocument doc) {
     String retval = "";
 
     try {
       final ByteArrayBuffer buf = new ByteArrayBuffer(doc.getDocLength().intValue());
+      byte[] blob;
       for (CmsDocumentBlobSegment seg : doc.getBlobSegments()) {
-        buf.append(seg.getDocBlob(), 0, seg.getDocBlob().length);
+        blob = seg.getDocBlob();
+        buf.append(blob, 0, blob.length);
       }
 
       final byte[] bytes = buf.buffer();
       LOGGER.debug("DAO: bytes len={}", bytes.length);
       retval = DatatypeConverter.printBase64Binary(bytes);
     } catch (Exception e) {
-      LOGGER.error("ERROR DECOMPRESSING PLAIN! {}", e.getMessage());
-      throw new ServiceException("ERROR DECOMPRESSING PLAINE! " + e.getMessage(), e);
+      errorDecompressing(e);
     }
 
     return retval;
@@ -272,8 +277,8 @@ public class CmsDocumentDao extends BaseDaoImpl<CmsDocument> {
    * @param doc LZW archive to decompress
    * @return base64-encoded String of decompressed document
    */
-  @SuppressFBWarnings("PATH_TRAVERSAL_IN") // There is no path traversal here
-  private String decompressLZW(CmsDocument doc) {
+  @SuppressFBWarnings("PATH_TRAVERSAL_IN") // No path traversal here
+  protected String decompressLZW(CmsDocument doc) {
     String retval = "";
 
     File src = null;
@@ -287,12 +292,17 @@ public class CmsDocumentDao extends BaseDaoImpl<CmsDocument> {
       errorDecompressing(e);
     }
 
+    if (src == null || tgt == null) {
+      throw new ServiceException("LZW DE-COMPRESSION ERROR: source and target cannot be null!");
+    }
+
     try {
       blobSegmentsToFile(doc, src);
-      // DECOMPRESS!
-      final LZWEncoder lzw = new LZWEncoder();
+
+      final LZWEncoder lzw = lzwSupplier.get();
       lzw.fileCopyUncompress(src.getAbsolutePath(), tgt.getAbsolutePath());
-      retval = DatatypeConverter.printBase64Binary(Files.readAllBytes(Paths.get(tgt.getAbsolutePath())));
+      retval =
+          DatatypeConverter.printBase64Binary(Files.readAllBytes(Paths.get(tgt.getAbsolutePath())));
 
       final boolean srcDeletedSuccessfully = src.delete();
       if (!srcDeletedSuccessfully) {
@@ -306,11 +316,12 @@ public class CmsDocumentDao extends BaseDaoImpl<CmsDocument> {
     } catch (Exception e) {
       errorDecompressing(e);
     }
+
     return retval;
   }
 
-  private void blobSegmentsToFile(CmsDocument doc, File src) {
-    try (FileOutputStream fos = new FileOutputStream(src);){
+  protected void blobSegmentsToFile(CmsDocument doc, File src) {
+    try (FileOutputStream fos = new FileOutputStream(src);) {
       for (CmsDocumentBlobSegment seg : doc.getBlobSegments()) {
         final byte[] bytes = seg.getDocBlob();
         fos.write(bytes);
@@ -321,10 +332,9 @@ public class CmsDocumentDao extends BaseDaoImpl<CmsDocument> {
     }
   }
 
-
   /**
-   * Compress (deflate) a document into LZW-compressed by calling native
-   * library and split on blob segments
+   * Compress (deflate) a document into LZW-compressed by calling native library and split on blob
+   * segments
    *
    * <p>
    * OPTION: Trap std::exception in shared library and return error code. The LZW library currently
@@ -338,11 +348,11 @@ public class CmsDocumentDao extends BaseDaoImpl<CmsDocument> {
    * </p>
    *
    * @param doc LZW archive to decompress
+   * @param base64 base64-encoded String
    * @return base64-encoded String of decompressed document
    */
-  @SuppressFBWarnings("PATH_TRAVERSAL_IN") // There is no path traversal here
-  private List<CmsDocumentBlobSegment> compressLZW(CmsDocument doc, String base64) {
-
+  @SuppressFBWarnings("PATH_TRAVERSAL_IN") // No path traversal here
+  protected List<CmsDocumentBlobSegment> compressLZW(CmsDocument doc, String base64) {
     final List<CmsDocumentBlobSegment> blobs = new ArrayList<>();
 
     File src = null;
@@ -356,12 +366,16 @@ public class CmsDocumentDao extends BaseDaoImpl<CmsDocument> {
       errorCompressing(e);
     }
 
+    if (src == null || tgt == null) {
+      throw new ServiceException("LZW COMPRESSION ERROR: source and target cannot be null!");
+    }
+
     try (FileOutputStream fos = new FileOutputStream(src);) {
       final byte[] bytes = DatatypeConverter.parseBase64Binary(base64.trim());
       fos.write(bytes, 0, bytes.length);
       fos.flush();
-      // COMPRESS!
-      final LZWEncoder lzw = new LZWEncoder();
+
+      final LZWEncoder lzw = lzwSupplier.get();
       lzw.fileCopyCompress(src.getAbsolutePath(), tgt.getAbsolutePath());
 
       final byte[] compressed = Files.readAllBytes(Paths.get(tgt.getAbsolutePath()));
@@ -381,7 +395,7 @@ public class CmsDocumentDao extends BaseDaoImpl<CmsDocument> {
         final String sequence = StringUtils.leftPad(String.valueOf(++i), 4, '0');
         final int segmentLength = min(compressed.length - segmentStart, BLOB_SEGMENT_LENGTH);
         blobs.add(new CmsDocumentBlobSegment(doc.getId(), sequence,
-                Arrays.copyOfRange(compressed,segmentStart, segmentStart + segmentLength)));
+            Arrays.copyOfRange(compressed, segmentStart, segmentStart + segmentLength)));
         segmentStart += segmentLength;
       }
 
@@ -400,14 +414,18 @@ public class CmsDocumentDao extends BaseDaoImpl<CmsDocument> {
     return blobs;
   }
 
-  private void errorDecompressing(Exception e) {
+  protected void errorDecompressing(Exception e) {
     LOGGER.error("ERROR DECOMPRESSING LZW! {}", e.getMessage(), e);
     throw new ServiceException("ERROR DECOMPRESSING LZW! " + e.getMessage(), e);
   }
 
-  private void errorCompressing(Exception e)  {
-    LOGGER.error("ERROR COMPRESSING LZW! {}", e.getMessage());
+  protected void errorCompressing(Exception e) {
+    LOGGER.error("ERROR COMPRESSING LZW! {}", e.getMessage(), e);
     throw new ServiceException("ERROR COMPRESSING LZW! " + e.getMessage(), e);
+  }
+
+  public void setLzwSupplier(Supplier<LZWEncoder> lzwSupplier) {
+    this.lzwSupplier = lzwSupplier;
   }
 
 }
